backend/requirements.txt
# Dépendances principales (Python 3.11+ recommandé, ok 3.13)
Flask>=3.0.0
Flask-Cors>=4.0.0
python-dotenv>=1.0.1
requests>=2.32.0
beautifulsoup4>=4.12.3
openai>=1.46.0
graphviz>=0.20.3
gunicorn>=21.2.0
networkx>=3.3
